requirements.txt
pinecone==6.0.0.dev3
sentence-transformers==2.2.2
torch==2.4.1  # CPU version for embeddings
streamlit==1.28.0
pypdf==3.17.1
groq
huggingface-hub==0.17.3
rapidfuzz==3.10.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic-settings==2.3.4
psutil==5.9.8
